+++ src/SyncAPIConnector/records/SpreadRecord.cs
@@ -6,19 +6,19 @@ namespace xAPI.Records;
 [DebuggerDisplay("{Symbol}, value:{Value}")]
 public record SpreadRecord : IBaseResponseRecord, ISymbol
 {
-    public long? Precision { get; set; }
+    public int? Precision { get; set; }
 
     public string? Symbol { get; set; }
 
     public long? Value { get; set; }
 
-    public long? QuoteId { get; set; }
+    public int? QuoteId { get; set; }
 
     public void FieldsFromJsonObject(JsonObject value)
     {
         Symbol = (string?)value["symbol"];
-        Precision = (long?)value["precision"];
+        Precision = (int?)value["precision"];
         Value = (long?)value["value"];
-        QuoteId = (long?)value["quoteId"];
+        QuoteId = (int?)value["quoteId"];
     }
 }
